DistinctBinaryTree
This is an example of the program running on the computer’s console. In
this example, values of n=3, n=4, and n=5 are used to find out how many
distinct binary trees are possible as well as a representation of the
trees.
EX: With n = 3. The program output corresponds to the binary tree as
follows: a) 123, b) 132, c) 213, d) 312, and e) 321.